pyproject.toml
[project]
name = "cash-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
]
